+++ 1.21/Common/src/main/java/fuzs/puzzleslib/api/client/gui/v2/components/tooltip/ClientComponentSplitter.java
@@ -1,6 +1,7 @@
 package fuzs.puzzleslib.api.client.gui.v2.components.tooltip;
 
 import net.minecraft.client.Minecraft;
+import net.minecraft.client.resources.language.ClientLanguage;
 import net.minecraft.network.chat.FormattedText;
 import net.minecraft.util.FormattedCharSequence;
 
@@ -18,24 +19,47 @@ public final class ClientComponentSplitter {
     }
 
     /**
-     * Split a formatted text instance according to a max width.
+     * Split formatted text instances according to a max width.
      *
-     * @param tooltipLines component to split for building tooltip
+     * @param tooltipLines components for building the tooltip
      * @return stream of split char sequences
      */
     public static Stream<FormattedCharSequence> splitTooltipLines(FormattedText... tooltipLines) {
-        return splitTooltipLines(Arrays.asList(tooltipLines));
+        return splitTooltipLines(170, Arrays.asList(tooltipLines));
     }
 
     /**
-     * Split a formatted text instance according to a max width.
+     * Split formatted text instances according to a max width.
      *
-     * @param tooltipLines components to split for building tooltip
+     * @param maxWidth     the maximum text width for the line splitter
+     * @param tooltipLines components for building the tooltip
+     * @return stream of split char sequences
+     */
+    public static Stream<FormattedCharSequence> splitTooltipLines(int maxWidth, FormattedText... tooltipLines) {
+        return splitTooltipLines(maxWidth, Arrays.asList(tooltipLines));
+    }
+
+    /**
+     * Split formatted text instances according to a max width.
+     *
+     * @param tooltipLines components for building the tooltip
      * @return stream of split char sequences
      */
     public static Stream<FormattedCharSequence> splitTooltipLines(List<? extends FormattedText> tooltipLines) {
+        return splitTooltipLines(170, tooltipLines);
+    }
+
+    /**
+     * Split formatted text instances according to a max width.
+     *
+     * @param maxWidth     the maximum text width for the line splitter
+     * @param tooltipLines components for building the tooltip
+     * @return stream of split char sequences
+     */
+    public static Stream<FormattedCharSequence> splitTooltipLines(int maxWidth,
+                                                                  List<? extends FormattedText> tooltipLines) {
         return tooltipLines.stream().flatMap((FormattedText formattedText) -> {
-            List<FormattedCharSequence> lines = Minecraft.getInstance().font.split(formattedText, 170);
+            List<FormattedCharSequence> lines = Minecraft.getInstance().font.split(formattedText, maxWidth);
             if (lines.isEmpty()) {
                 // empty components yield an empty list
                 // since empty lines are desired on tooltips make sure they don't go missing
@@ -43,6 +67,28 @@ public final class ClientComponentSplitter {
             } else {
                 return lines.stream();
             }
+        });
+    }
+
+    /**
+     * Process formatted text instances into char sequences.
+     *
+     * @param tooltipLines components for building the tooltip
+     * @return stream of char sequences
+     */
+    public static Stream<FormattedCharSequence> processTooltipLines(FormattedText... tooltipLines) {
+        return processTooltipLines(Arrays.asList(tooltipLines));
+    }
+
+    /**
+     * Process formatted text instances into char sequences.
+     *
+     * @param tooltipLines components for building the tooltip
+     * @return stream of char sequences
+     */
+    public static Stream<FormattedCharSequence> processTooltipLines(List<? extends FormattedText> tooltipLines) {
+        return tooltipLines.stream().map((FormattedText formattedText) -> {
+            return ClientLanguage.getInstance().getVisualOrder(formattedText);
         });
     }
 }