+++ 1.21/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.0.16-1.21] - 2024-07-15
+### Changed
+- Allow `TooltipComponent` to optionally not split tooltip lines
+- Minor additions to `ClientComponentSplitter`
+
 ## [v21.0.15-1.21] - 2024-07-14
 ### Added
 - Add `ItemHelper`